The functional volume push tests actually work, via some terrible hacks.

+++ flocker/volume/filesystems/zfs.py
@@ -149,7 +149,9 @@ class Filesystem(object):
             process.stdin.close()
             succeeded = not process.wait()
         if succeeded:
-            subprocess.check_call([b"zfs", b"destroy", b"-R", self.name])
+            exists = not subprocess.Popen([b"zfs", b"list", self.name]).wait()
+            if exists:
+                subprocess.check_call([b"zfs", b"destroy", b"-R", self.name])
             subprocess.check_call([b"zfs", b"rename", temp_filesystem,
                                    self.name])
             subprocess.check_call([b"zfs", b"set",
